Add Sarthaki to the contributors list
Hi! I'm Sarthaki 👋
This pull request adds my name and GitHub username (@Sarthaki2007) to the `contribution.md` file under the Contributors section, as requested in Issue #1.
This is my first contribution to Alum-X-Frontend, and I’m really excited to be part of this project and the open-source community. Looking forward to learning and contributing more in the future!
Thank you for this opportunity 🚀